Move serde into its own feature
