+++ Cargo.toml
@@ -59,9 +59,9 @@ png = { version = "0.17", optional = true }
 # Compression/decompression
 flate2 = "1.0"
 # Serialization/Deserialization
-serde = { version = "1.0", features = ["derive"] } # MOM serialization/deserialization
-bincode = "1.3.3"
-toml = "0.8"
+serde = { version = "1.0", features = ["derive"], optional = true } # MOM serialization/deserialization
+bincode = { version = "1.3.3", optional = true }
+toml = { version = "0.8", optional = true }
 # MOM specific
 itertools = "0.14"
 # For date in FITS written files
@@ -84,8 +84,10 @@ harness = false
 
 [features]
 base64 = ["dep:base64"]
-skymap = ["dep:colorous", "dep:mapproj", "dep:png"]
 memmap = ["dep:memmap2"]
+skymap = ["dep:colorous", "dep:mapproj", "dep:png"]
+serde = ["dep:serde"]
+sort = ["dep:bincode", "dep:toml", "memmap", "serde"]
 
 # See opti here: https://github.com/johnthagen/min-sized-rust
 [profile.release]
