+++ crates/cli/Cargo.toml
@@ -19,7 +19,7 @@ bench = false
 test = false
 
 [dependencies]
-hpxlib = { package = "cdshealpix", path = "../..", features = ["memmap", "skymap"] }
+hpxlib = { package = "cdshealpix", path = "../..", features = ["skymap", "sort"] }
 log = { workspace = true }
 mapproj = { workspace = true }
 memmap2 = { workspace = true }
