+++ src/nested/map/mom/impls/ported_from_mom_builder/state/value/value_state.rs
@@ -1,7 +1,9 @@
+#[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
 /// Simple leaf state with a single value.
-#[derive(Clone, Copy, Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Copy, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct ValueState<T>(pub T);
 
 macro_rules! impl_from_value_state {
